Update SongStartTimePacket also long now from float

+++ BeatTogether.DedicatedServer.Kernel/Managers/Abstractions/IGameplayManager.cs
@@ -12,7 +12,7 @@ namespace BeatTogether.DedicatedServer.Kernel.Managers.Abstractions
         GameplayManagerState State { get; }
 		BeatmapIdentifier? CurrentBeatmap { get; }
 		GameplayModifiers CurrentModifiers { get; }
-        public float _songStartTime { get; }
+        public long _songStartTime { get; }
 
         void HandlePlayerLeaveGameplay(IPlayer player);
         void HandleGameSceneLoaded(IPlayer player, SetGameplaySceneReadyPacket packet);
